Puntos 2, 3, 4, 5, 6, 7 completos.

+++ tp1_ValleAlejandro/src/ar/edu/unju/fi/ejercicio1/Main.java
@@ -4,7 +4,7 @@ public class Main {
 
 	public static void main(String[] args) {
 		int a, b, c, d, e;
-		float suma, resultado;
+		double suma, resultado;
 		a = 8;
 		b = 5;
 		c = 1;
